Per http://jira.codehaus.org/secure/ViewIssue.jspa?key=PRV-21 ( SnapshotManager should be an interface )
Made SnaphotManager into an interface and added the following classes...

AbstractBaseSnaphotManager - provides 99% of the implementation for any given SnapshotManager implementation. All the following classes extend this class (or another class that extends this class).

JavaSnaphotManager - What SnapshotManager used to be. Simply uses standard Java Serialization for snaphots

SkaringaSnapshotManager - What XmlSnaphotManager used to be

XStreamSnaphotManager - Adds XStream XML snapshot management. This is an alternative to Skaringa XML snapshot management.


NullSnapshotManager now extends JavaSnapshotManager, but is otherwise unchanged. As mentioned previously, XmlSnapshotManager is to be removed and replaced by SkaringaSnapshotManager.

Also added xstream.jar + license to root/lib and added two Main classes in the demo2 to provide separate examples for Skaringa and Xstream snapshotting.  Updated build files to provide access to new demo2 Main classes and added an XSLT stylesheet to provide a nice view of the build file in the browser.

Jake